refactor: cleaning FakeProviderSyntaxRewriter part 4 #85 #89
